requirements.txt
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
cryptography==39.0.1
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
geographiclib==2.0
geopy==2.3.0
ipykernel==6.21.1
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
plotly==5.13.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
six==1.16.0
stack-data==0.6.2
tenacity==8.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
Werkzeug==2.2.2
